+++ src/Profile/index.js
@@ -1,13 +1,10 @@
 import React from "react";
-import CommonUsers from "./CommonUsers";
 import Info from "./Info";
 
 export default function() {
   return (
     <div>
       <Info />
-
-      <CommonUsers />
     </div>
   );
 }
