+++ src/Suggestions/icons/delete.png
